+++ src/HTTP/Client.php
@@ -61,8 +61,7 @@ interface Client
     /**
      * send a request and return the response
      *
-     * @param Client::METHOD_GET $method
-     * @param string $url
+     * @param Client::METHOD_* $method
      * @param string[] $headers
      *
      * @throws HttpException if anything goes wrong requesting the data